Begin rewrite of SIS scraper
With the switch to a new service for SIS on 09/17/2025, we decided to rewrite our existing SIS scraper in favor of simpler logic and better maintainability.

The old scrapers still work, we just don't really like touching that code.

This new scraper IS NOT COMPLETE yet. I discovered a working procedure via Postman for getting all course data, and am in the middle of automating it using Python.

A current issue is that the GET request made in get_cookie() does not return a cookie, which is needed to access any course data. This is actively being looked into.