Redefine "class" and "course" terminology
In the context of this scraper, a "class" refers to a section of a course, while a "course" refers to the overarching course that may have multiple classes.

The data returned from SIS is keyed by classes, not courses. The get_course_data() function manipulates and aggregates this data such that the returned structure is keyed by courses instead, with classes as a sub-field of each course.