Refine asynchronous logic sections
With a bit of help from Claude Sonnet 4, the scraper has been further optimized such that data is scraped from every subject within a term in parallel, each one using its own client session to avoid session state conflicts in the SIS backend.

There are currently still issues with the scraping logic as too many assumptions are being made about the scraped data.

For example: after a couple of runs of the scraper, it turns out that not every faculty member has a non-null email address. This is going to require some investigating into the SIS data.