Add get_reverse_subject_map()
This was sparked by an issue in which STSS 2520 appears as a prerequisite for some other course in Fall 2025, but STSS no longer exists. However, we still want to convert the subject name into its respective code, so a function that could create a "master" subject reverse lookup map across multiple terms was needed.