Remove old SIS scraper code
The new scraper code should be about the same or better than the old code at this point, and there isn't much of a point in keeping it as a reference anymore.

If needed, the code should still exist in the CARPI Data repository, which will probably soon be archived.