Make subject name-code map optional in prerequisite parsing
I edited a little bit of code such that the get_class_prerequisites() will attempt to map subject names to codes if a mapping is provided, but will otherwise leave it alone. This behavior is consistent with get_class_corequisites() and get_class_crosslists().

I also commented out a simple check in prereq_parser.py that enforces subject codes.